+++ src/main/java/co/edu/ucentral/repository/IComercioRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import co.edu.ucentral.model.Comercio;
 
 public interface IComercioRepository extends JpaRepository<Comercio, Integer> {
-	Comercio findByDocumento(int documento);
-	boolean  existsByDocumento(int documento);	
+	Comercio findByDocumento(long documento);
+	boolean  existsByDocumento(long documento);	
 }
